+++ pkg/props/proportion.go
@@ -1,0 +1,9 @@
+package props
+
+// Proportion represents a proportion from a rectangle, example: 16x9, 4x3...
+type Proportion struct {
+	// Width from the rectangle: Barcode, image and etc.
+	Width float64
+	// Height from the rectangle: Barcode, image and etc.
+	Height float64
+}
